implemented global feature

+++ server/controllers/recipe-controller.js
@@ -5,6 +5,19 @@ const { User, Recipe } = require('../models')
 const { AuthenticationError } = require('apollo-server-express')
 
 const recipeController = {
+    async getRecipes(parent, params, context) {
+        if (context.user) {
+            try {
+                const recipes = await Recipe.find().sort({ createdAt: -1})
+                return recipes
+            } catch (e) {
+                console.error(e)
+                throw new AuthenticationError('Unable to return recipes')
+            }
+        } 
+        
+        throw new AuthenticationError('You must be logged in')
+    },
     async createRecipe(parent, { type, name, brand, ...recipe}, context) {
         if (context.user) {
             try {
@@ -29,11 +42,17 @@ const recipeController = {
             }
         }
 
-        throw new AuthenticationError('You must be logged in')
+        
     },
     async removeRecipe(parent, { id }, context) {
         if (context.user) {
             try {
+
+                const user = await User.findById(context.user._id)
+                if (!user.recipes.some(recipeID => recipeID.toString() === id)) {
+                    throw new AuthenticationError('You are not autheticated to delete this recipe')
+                }
+
                 const deletedRecipe = await Recipe.findByIdAndDelete(id)
 
                 const updatedUser = await User.findByIdAndUpdate(
